requirements.win.txt
# Windows-specific requirements with version constraints to avoid build issues
# This file includes tensorflow-cpu for local Windows development

# Core dependencies (same as main requirements.txt)
gradio
requests
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-huggingface
langchain-groq
langchain-tavily
langchain-chroma
langgraph
huggingface_hub
supabase
arxiv
pymupdf
wikipedia
pgvector
python-dotenv
pytesseract
matplotlib
pandas
numpy
pillow
jupyter
openpyxl
beautifulsoup4
lxml
sentence-transformers
google-genai
litellm
board_to_fen
board-to-fen
scipy
scikit-learn
sympy
networkx
nltk
opencv-python

# Windows-specific: TensorFlow for local development
# tensorflow-cpu
# Alternative options if tensorflow-cpu fails:
# tensorflow
# tensorflow-cpu==2.15.0
# tensorflow==2.15.0 